build using cmake.

+++ dep/tinyobjloader/tiny_obj_loader.cpp
@@ -1,6 +1,6 @@
 //
 // Copyright 2012-2013, Syoyo Fujita.
-// 
+//
 // Licensed under 2-clause BSD liecense.
 //
 
@@ -62,7 +62,7 @@ static inline bool isNewLine(const char c) {
   return (c == '\r') || (c == '\n') || (c == '\0');
 }
 
-// Make index zero-base, and also support relative index. 
+// Make index zero-base, and also support relative index.
 static inline int fixIndex(int idx, int n)
 {
   int i;
@@ -145,7 +145,7 @@ static vertex_index parseTriple(
       token += strcspn(token, "/ \t\r");
       return vi;
     }
-    
+
     // i/j/k or i/j
     vi.vt_idx = fixIndex(atoi(token), vtsize);
     token += strcspn(token, "/ \t\r");
@@ -157,7 +157,7 @@ static vertex_index parseTriple(
     token++;  // skip '/'
     vi.vn_idx = fixIndex(atoi(token), vnsize);
     token += strcspn(token, "/ \t\r");
-    return vi; 
+    return vi;
 }
 
 static unsigned int
@@ -178,7 +178,7 @@ updateVertex(
     return it->second;
   }
 
-  assert(in_positions.size() > (3*i.v_idx+2));
+  assert(in_positions.size() > (unsigned int)(3*i.v_idx+2));
 
   positions.push_back(in_positions[3*i.v_idx+0]);
   positions.push_back(in_positions[3*i.v_idx+1]);
@@ -299,7 +299,7 @@ std::string LoadMtl (
   std::stringstream err;
 
   material_t material;
-  
+
   int maxchars = 8192;  // Alloc enough size.
   std::vector<char> buf(maxchars);  // Alloc enough size.
   while (inStream.peek() != -1) {
@@ -326,9 +326,9 @@ std::string LoadMtl (
 
     assert(token);
     if (token[0] == '\0') continue; // empty line
-    
+
     if (token[0] == '#') continue;  // comment line
-    
+
     // new mtl
     if ((0 == strncmp(token, "newmtl", 6)) && isSpace((token[6]))) {
       // flush previous material.
@@ -344,7 +344,7 @@ std::string LoadMtl (
       material.name = namebuf;
       continue;
     }
-    
+
     // ambient
     if (token[0] == 'K' && token[1] == 'a' && isSpace((token[2]))) {
       token += 2;
@@ -355,7 +355,7 @@ std::string LoadMtl (
       material.ambient[2] = b;
       continue;
     }
-    
+
     // diffuse
     if (token[0] == 'K' && token[1] == 'd' && isSpace((token[2]))) {
       token += 2;
@@ -366,7 +366,7 @@ std::string LoadMtl (
       material.diffuse[2] = b;
       continue;
     }
-    
+
     // specular
     if (token[0] == 'K' && token[1] == 's' && isSpace((token[2]))) {
       token += 2;
@@ -377,7 +377,7 @@ std::string LoadMtl (
       material.specular[2] = b;
       continue;
     }
-    
+
     // transmittance
     if (token[0] == 'K' && token[1] == 't' && isSpace((token[2]))) {
       token += 2;
@@ -517,7 +517,7 @@ LoadObj(
     basePath = mtl_basepath;
   }
   MaterialFileReader matFileReader( basePath );
-  
+
   return LoadObj(shapes, ifs, matFileReader);
 }
 
@@ -565,7 +565,7 @@ std::string LoadObj(
 
     assert(token);
     if (token[0] == '\0') continue; // empty line
-    
+
     if (token[0] == '#') continue;  // comment line
 
     // vertex
@@ -614,7 +614,7 @@ std::string LoadObj(
       }
 
       faceGroup.push_back(face);
-      
+
       continue;
     }
 
@@ -641,13 +641,13 @@ std::string LoadObj(
       char namebuf[4096];
       token += 7;
       sscanf(token, "%s", namebuf);
-        
+
       std::string err_mtl = readMatFn(namebuf, material_map);
       if (!err_mtl.empty()) {
         faceGroup.clear();  // for safety
         return err_mtl;
       }
-      
+
       continue;
     }
